+++ app/helpers/additionals_icons_helper.rb
@@ -41,21 +41,7 @@ module AdditionalsIconsHelper
   end
 
   def additionals_asset_path(file)
-    plugin_id = 'additionals'
-
-    if Additionals.redmine6?
-      asset_path "plugin_assets/#{plugin_id}/#{file}"
-    else
-      "#{additionals_image_path plugin_id}/#{file}"
-    end
-  end
-
-  def additionals_image_path(plugin_id)
-    return asset_path if Additionals.redmine6?
-    return @additionals_image_path if defined? @additionals_image_path
-
-    relative_url = Redmine::Utils.relative_url_root
-    @additionals_image_path = "#{relative_url}/plugin_assets/#{plugin_id}/images"
+    asset_path "plugin_assets/additionals/#{file}"
   end
 
   private