+++ lib/additionals/plugin_version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Additionals
-  module PluginVersion
-    VERSION = '4.0.0-main' unless defined? Additionals::PluginVersion::VERSION
-  end
-end