+++ test/test_helper.rb
@@ -37,24 +37,8 @@ module Additionals
       end
     end
 
-    def smiley_test_span(text)
-      "<span class=\"additionals smiley a-icon smiley-smiley\" title=\":)\">#{text}</span>"
-    end
-
-    def svg_test_icon
-      '<svg class="a-s18 a-svg-icon smiley" title=":)" aria-hidden="true">' \
-        '<use href="/plugin_assets/additionals/images/icons.svg#icon--smiley-smiley"></use></svg>'
-    end
-
-    # for tanuki_emoji 0.11.0 or newer
-    # NOTE: tanuki_emoji 0.11.0 can only be used, if Redmine 5.0 suppord dropped
-    # (depenceny for i18n version)
-    def emoji_heart_tag_new
-      '<additionals-emoji title="red heart" data-name="red_heart" data-unicode-version="6.0">❤️</additionals-emoji>'
-    end
-
     def emoji_heart_tag
-      '<additionals-emoji title="heavy black heart" data-name="heart" data-unicode-version="1.1">❤</additionals-emoji>'
+      '<additionals-emoji title="red heart" data-name="heart" data-unicode-version="6.0">❤️</additionals-emoji>'
     end
   end