+++ test/unit/project_test.rb
@@ -169,10 +169,6 @@ class ProjectTest < Additionals::TestCase
   def test_available_status_ids
     ids = Project.available_status_ids
 
-    if Redmine::VERSION.to_s < '5.1'
-      assert_equal 3, ids.count
-    else
-      assert_operator ids.count, :>, 3
-    end
+    assert_operator ids.count, :>, 3
   end
 end