+++ test/unit/wiki_formating/common_mark_test.rb
@@ -13,13 +13,9 @@ module WikiFormatting
     def test_smileys
       with_plugin_settings 'additionals', legacy_smiley_support: 1,
                                           emoji_support: 0 do
-        input = <<~HTML
-          A small test :) with an smiley
-        HTML
-        expected = <<~HTML
-          A small test #{smiley_test_span svg_test_icon} with an smiley
-        HTML
-        assert_equal expected, smiley_filter(input)
+
+        assert smiley_filter('A small test :) with an smiley').include?('#icon--smiley-smiley')
+        assert smiley_filter('A small test :) with an smiley').include?('a-s18 a-svg-icon smiley')
       end
     end